move util function out, more useEffect up
