+++ src/base/asField.jsx
@@ -4,12 +4,7 @@ import { getValidateText } from './validate';
 import isDeepEqual from 'deep-equal';
 import { usePrevious } from '../hooks';
 import { isHidden, isDependShow } from './isHidden';
-import {
-  evaluateString,
-  isLooselyNumber,
-  isCssLength,
-  isFunction,
-} from './utils';
+import { isLooselyNumber, isCssLength, convertValue } from './utils';
 
 // asField拆分成逻辑组件和展示组件，从而可替换展示组件的方式完全插拔fr的样式
 export const asField = ({ FieldUI, Widget }) => {
@@ -40,26 +35,23 @@ export const asField = ({ FieldUI, Widget }) => {
     } = rest;
     const prevValue = usePrevious(_value);
     // most key of schema, disabled, readonly, options, hidden, support for function expression
-    const convertValue = item => {
-      if (typeof item === 'function') {
-        return item(formData, rootValue);
-      } else if (typeof item === 'string' && isFunction(item) !== false) {
-        const _item = isFunction(item);
-        try {
-          return evaluateString(_item, formData, rootValue);
-        } catch (error) {
-          console.error(error.message);
-          console.error(`happen at ${item}`);
-          return item;
-        }
+    useEffect(() => {
+      // 首次渲染不做
+      if (firstRender.current) {
+        firstRender.current = false;
+        return;
       }
-      return item;
-    };
+      // 之后每次改动就算touch了
+      if (isDeepEqual(prevValue, rest.value)) {
+      } else {
+        fieldTouched.current = true;
+      }
+    }, [rest.value]);
 
-    hidden = convertValue(hidden);
-    disabled = convertValue(disabled);
-    readonly = convertValue(readonly);
-    options = convertValue(options);
+    hidden = convertValue(hidden, formData, rootValue);
+    disabled = convertValue(disabled, formData, rootValue);
+    readonly = convertValue(readonly, formData, rootValue);
+    options = convertValue(options, formData, rootValue);
     // iterate over schema, and convert every key
     let _schema = { ...schema };
     Object.keys(schema).forEach(key => {
@@ -77,7 +69,7 @@ export const asField = ({ FieldUI, Widget }) => {
       ];
       // TODO: need to cover more
       if (availableKey.indexOf(key) > -1) {
-        _schema[key] = convertValue(schema[key]);
+        _schema[key] = convertValue(schema[key], formData, rootValue);
       }
     });
 
@@ -147,19 +139,6 @@ export const asField = ({ FieldUI, Widget }) => {
       validateText,
       labelWidth,
     };
-
-    useEffect(() => {
-      // 首次渲染不做
-      if (firstRender.current) {
-        firstRender.current = false;
-        return;
-      }
-      // 之后每次改动就算touch了
-      if (isDeepEqual(prevValue, rest.value)) {
-      } else {
-        fieldTouched.current = true;
-      }
-    }, [rest.value]);
 
     return (
       <FieldUI {...fieldProps}>
